==> Scripts/05_2_MakePredictions.ipynb <==
import catboost as cb
import pandas as pd
import os
import random
random.seed(1667)
os.chdir("C:/Users/dread/Downloads/educationopportunity")

#load data and models
predframe = pd.read_csv("Data/final_preprediction_frames/total_prepred.csv")
total = cb.CatBoostRegressor()
total.load_model("Models/catboost")
math = cb.CatBoostRegressor()
math.load_model("Models/catboostMath")
erw = cb.CatBoostRegressor()
erw.load_model("Models/catboostERW")

#predict
SAT_Total = total.predict(predframe[['exposure_parent_ses_hs','ec_high_parent_ses_hs','volunteering_rate_hs','ec_high_own_ses_hs','ec_parent_ses_hs','nonwhite_share2010','exposure_own_ses_hs','black','frac_coll_plus2016','ec_own_ses_hs','white']])
SAT_Math = math.predict(predframe[['exposure_parent_ses_hs','ec_high_parent_ses_hs','volunteering_rate_hs','ec_high_own_ses_hs','ec_parent_ses_hs','nonwhite_share2010','exposure_own_ses_hs','black','frac_coll_plus2016','ec_own_ses_hs','white']])
SAT_ERW = erw.predict(predframe[['exposure_parent_ses_hs', 'volunteering_rate_hs', 'ec_high_parent_ses_hs', 'ec_high_own_ses_hs', 'nonwhite_share2010', 'ec_parent_ses_hs', 'exposure_own_ses_hs', 'white', 'frac_coll_plus2016', 'civic_organizations_zip', 'black', 'hispanic']])

predframe['SAT_Total'] = SAT_Total
predframe['SAT_Math'] = SAT_Math
predframe['SAT_ERW'] = SAT_ERW

#save
predframe.to_csv("Data/final_postprediction_frames/total_postpred.csv")
